finished the five features (add,showall,showitem,update,delete)

+++ Servlet-JSP-Project/Items-Web-Application/src/main/java/service/ItemService.java
@@ -12,7 +12,7 @@ public interface ItemService {
 	
 	List<Item> getAllItems();
 	
-	Item selectItem(Long id);
+	Item selectItem(Integer id);
 	
 	boolean saveItem(Item item);
